+++ src/main.ts
@@ -3,7 +3,7 @@ import { tokenize } from './lexer/tokenizer';
 import { makeSyntaxTree } from './lexer/parser';
 import { runtime } from './runtime';
 import { isLeft } from 'fp-ts/lib/Either';
-import { render } from './renderer';
+import { render } from './abc-renderer';
 
 function main() {
     let args = process.argv;
